Estrutura base, com componentes na pasta componetes, para que possamos todos trabalhar
