--- app/src/main/java/com/tesp/tindogapp/pages/janelaBase.kt
@@ -1,113 +1,0 @@
-package com.tesp.tindogapp.pages
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            TinDogTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    Page()
-                }
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true, heightDp = 600, widthDp = 380)
-@Composable
-fun Page(): Unit {
-    Box(
-        modifier = Modifier
-            .padding(0.dp) // Set the margins here
-    ) {
-
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp)
-        ) {
-            Logotipo()
-            CaixaImputs ()
-            // Logotipo
-            // box salmao recebe imputs
-            //imputs
-
-        }
-    }
-}
-//@Preview ()
-@Composable
-fun Logotipo () {
-    Box(contentAlignment = Alignment.Center,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(250.dp)
-            .padding(20.dp)
-
-    ){
-        Image(
-            painter = painterResource(id = R.drawable.tindog_logo),
-            contentDescription = null
-        )
-    }
-}
-
-@Composable
-fun CaixaImputs() {
-    Box(
-        modifier = Modifier
-            .background(
-                Color(0xFFFFDBD2),
-                shape = RoundedCornerShape(16.dp)
-            )
-            .fillMaxSize()
-            .padding(0.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            // verticalArrangement = Arrangement.Center
-        ) {
-            Text(
-                text = "Login to your Dog House",
-                fontFamily = FontFamily.Monospace,
-                textAlign = TextAlign.Center,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 16.dp)
-            )
-        }
-    }
-}
-
